_requirements.txt
streamlit
google-generativeai
python-dotenv
pytesseract
tesseract
pdf2image
opencv-python
